# backend/docs/swagger.yaml
definitions:
...
  entities.JWT:
    properties:
      jwt:
        type: string
    type: object
...
  entities.RetFailed:
    properties:
      error:
        type: string
      msg:
        type: string
      status:
        type: integer
    type: object
...
  entities.RetSuccess-entities_JWT:
    properties:
      error:
        type: string
      msg:
        $ref: '#/definitions/entities.JWT'
      status:
        type: integer
    type: object
...
  entities.RetSuccess-string:
    properties:
      error:
        type: string
      msg:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: A place to document what I've learned.
  title: Coding Notes
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: login to get jwt token
      parameters:
      - description: user credentials
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RetSuccess-entities_JWT'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "412":
          description: Precondition Failed
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.RetFailed'
      summary: Login
      tags:
      - users
  /logout:
    post:
      consumes:
      - application/json
      description: Checks if jwt is valid
      parameters:
      - description: jwt token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.RetSuccess-string'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "412":
          description: Precondition Failed
          schema:
            $ref: '#/definitions/entities.RetFailed'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.RetFailed'
      summary: AuthorizeCheck
      tags:
      - users
...
swagger: "2.0"